registry/recipes/llamaindex-starter/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
llama-index-core==0.12.52
llama-index-llms-ollama==0.5.2
llama-index-embeddings-ollama==0.5.0
llama-index-vector-stores-qdrant==0.6.1
qdrant-client==1.12.1
pydantic==2.9.2
